+++ database/DDL/CREATE_wkf_appended.sql
@@ -7,7 +7,7 @@ CREATE  TABLE wkf_appended
  ( 
   id serial
  ,company_code VARCHAR (10)  NOT NULL
- ,application_number INT NOT NULL
+ ,application_number BIGINT NOT NULL
  ,approval_number serial
  ,approval_company_code VARCHAR(10) NOT NULL
  ,approval_unit_code VARCHAR(10) NOT NULL
